backend/science/requirements.txt
# Science Module Dependencies
# Owner: Science/ML Team

# LangGraph and LangChain
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0

# LLM Providers
langchain-google-genai>=0.0.5
langchain-openai>=0.0.5

# Configuration
python-dotenv>=1.0.0

# Core dependencies
pydantic>=2.0.0